pyproject.toml
[tool.poetry]
name = "gl-cli"
version = "0.1.0"
description = ""
authors = ["staticf0x <44530786+staticf0x@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gl_cli"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.1"
python-gitlab = "^3.13.0"
click = "^8.1.3"
rich = "^13.3.1"
arrow = "^1.2.3"


[tool.poetry.group.dev.dependencies]
devtools = {extras = ["pygments"], version = "^0.10.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
